feat: User delete endpoint

+++ src/api/users/users.controller.ts
@@ -1,9 +1,23 @@
-import { Body, Controller, Delete, Get, Param, Post, Put, Query, UseGuards, UseInterceptors } from '@nestjs/common';
+import {
+  BadRequestException,
+  Body,
+  Controller,
+  Delete,
+  Get,
+  Param,
+  Post,
+  Put,
+  Query,
+  UseGuards,
+  UseInterceptors
+} from '@nestjs/common';
 import { ApiTags } from '@nestjs/swagger';
 import { MongooseQueryParser } from 'mongoose-query-parser';
 import { User } from '../../database/models/User/User';
 import { HttpCacheInterceptor } from '../../utils/interceptors/http-cache.interceptor';
+import { UserData } from '../auth/decorators';
 import { AccessTokenGuard } from '../auth/guards';
+import { IUserData } from '../auth/interfaces/UserData';
 import { UsersService } from './users.service';
 
 
@@ -54,13 +68,19 @@ export class UsersController {
   /**
    * Endpoint to delete one record from mongoDB Database
    * @param id
+   * @param userData
    */
   @UseGuards(AccessTokenGuard)
   @Delete(':id')
   public async deleteUser(
-    @Param('id') id: string
+    @Param('id') id: string,
+    @UserData() userData: IUserData
   ) {
 
+    /** Check if the role permit this operation */
+    if (userData.roleLevel < 90) {
+      throw new BadRequestException('The user does not have permission to delete a user');
+    }
     return this.usersService.deleteUser(id);
 
   }
